BaseClass is FINALLY DONE!
After years of struggling, the Base of it all, is now done. I can now sleep. (dramatization)

= Changelog
- BaseClass is now basically finished, all that it is missing is animation.
- MapService no longer creates random IDs, since you'll only have one map per server anyways.
- MonsterService has some new snazzy functions to make it easier to fetch resources for the Monsters.
- Fixed inheritance problems

There may still be some remnants of bugfixing and tests. And the Attack function in BaseClass is still dependent on abstract numbers instead of Animation Events. Though BaseClass is now very much ready to use! :D